minor refactoring: move *Ops classes into subpackage
